implement ability for validators to return "rich" results instead of raising exceptions for errors
This is the first visible step for a major change in pycerberus: It turned
out that raising exceptions on validation errors has significant disadvantages.
In some scenarios I would like to access the converted/validated values even
if some values are invalid. Also it is a bit difficult to return multiple errors
from a single validator when using exceptions.

Initially (years ago) I assumed that using exceptions for validation errors
were a good thing in a validation library (despite the general advice to avoid
exceptions as control flow) so no error will be missed by accident.

However I came to the conclusion that plain result values are safe anyway
within pycerberus as callers will be aware of validation errors (that's the
point of a validation library, isn't it :-) and "rich result instances" will
enable a lot of important functionality which to present more helpful
validation results to users.

This commit just adds additional capabilities for Validator but no validator
currently uses them. Also Schema will benefit greatly from that but so far
there are no changes in that direction.

Validators can opt-in to the new scheme but the default is still to use
exceptions as before (I plan to swich the defaults for 1.0).