ability to mark an Error as "critical"
"Critical" errors are by definition more severe so some
checks might have been skipped (this happens for example if a
conversion to a Python type fails as following checks likely
depend on the type conversion).

Currently this is only used in a few places and mostly pycerberus
does not care about critical and non-critical errors now but it is
an important step for code higher in the stack to determine if
an error can be ignored safely.