ValidationTest: change default for .process() to ensure that the result does not contain errors
Previously (with the exception-based API) errors would be visible very
easily but "rich results" might hide the error if the caller does not
explicitely checks the result.

That seems to dangerous (as test code might not be reviewed that
often) so add a new parameter so the method DOES NOT bail out if
the result contains errors.

Basically this change ensures we have the same behavior with in the
"rich results" API as with the "exception-based" API.