SchemaValidator: use ".nr_errors()" instead of comparing lists so this also works with mutable data types
Currently ".errors" is always immutable but when reviewing the code I stumbled across
these lines and I had to check just to be sure. Hopefully this change removes doubts.

The slightly changed semantics (e.g. previously it aborted if a form validator removed
or replaced an error without changing the number) do not concern me as these seem to
be absolute edge cases and users can always work around that by subclassing/forking.