create deprecation warning when a validator does not set "exception_if_invalid" explicitely
Originally I copied FormEncode's API design but over the years I noticed
that more complex setups are harder to implement when using exceptions:
For example returning multiple errors from the same validator or ignoring
certain errors is hard if the validator raises an exception.

Right now validators can opt in to return "rich results" (aka Result instances)
and after several years of production use this has proven to be
usable API design so this should become the default in the future.

- In pycerberus 0.7 you will see deprecation warnings if a validator does
  not declare its preferred exception scheme.
- pycerberus 0.8 will switch the default for validators without explicit
  declarations and issue a deprecation warning if exceptions are used.
- Still I expect that raising InvalidData errors will be supported for
  quite a while, at least until version 1.0.